Refactor experiment helper classes
