+++ tests/test_cluster_factors/test_cluster_binary_factors.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 import yaml
 import pandas as pd
-from big_dataset_utility.cluster_binary_factors import (cluster_factors_corr, cluster_factors_voting, cluster_factors_rand,
+from tests.big_dataset_utility.feature_labeling.cluster_binary_factors import (cluster_factors_corr, cluster_factors_voting, cluster_factors_rand,
                                                         cluster_factors_rand_un)
 
 # Define a fixture to load test cases from the YAML file
